+++ README.rst
@@ -54,7 +54,7 @@ Features
 - Android compatible
 - Support for JP Morgan's Quorum via `web3j-quorum <https://github.com/web3j/quorum>`_
 - Support for `EEA Privacy features as described in EEA documentation <https://entethalliance.org/technical-documents/>`_
-  and implemented in `Pegasys Pantheon <https://docs.pantheon.pegasys.tech/en/stable/Reference/Pantheon-API-Methods/#eea-methods>`_.
+  and implemented in `Pegasys Pantheon <https://docs.besu.pegasys.tech/en/stable/Reference/Pantheon-API-Methods/#eea-methods>`_.
 
 
 It has five runtime dependencies:
