+++ besu/src/main/java/org/web3j/protocol/besu/response/privacy/PrivGetTransactionReceipt.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2019 Web3 Labs LTD.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package org.web3j.protocol.besu.response.privacy;
+
+import java.util.Optional;
+
+import org.web3j.protocol.core.Response;
+
+public class PrivGetTransactionReceipt extends Response<PrivateTransactionReceipt> {
+    public Optional<PrivateTransactionReceipt> getTransactionReceipt() {
+        return Optional.ofNullable(getResult());
+    }
+}
