+++ besu/src/main/java/org/web3j/utils/PrivacyGroupUtils.java
