+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'base'
 
 include 'abi'
+include 'besu'
 include 'codegen'
 include 'console'
 include 'contracts'
@@ -10,7 +11,6 @@ include 'eea'
 include 'geth'
 include 'hosted-providers'
 include 'integration-tests'
-include 'pantheon'
 include 'parity'
 include 'rlp'
 include 'tuples'
